Initial commit
This is housekeeper V2, now in Go :)

This intial commit allows listing of all instances in any account
with the brkt-HouseKeeper role.

It will only be implemented for AWS for now, but is designed in a
way such that adding more CSPs in the future shouldn't be any
problem.

Jira: None