Get accounts from MV repo
This adds a small python script, which will fetch the list of AWS
accounts and who they belong to from the MV repo. This is saved
as a JSON file and passed to the Go program to be used.

A Dockerfile and Makefile was created to make this easier.

Jira: None