Retry deletion of snaphots on error
If we hit a RequestLimitExceeded error when trying to delete
snaphots in AWS, we should now retry — with exponential backoff.
The backoff implementation is really simple and stupid, but should
work for this case.

I'll extend this to the other resource types if this works, I need
to confirm this first though.

Change-Id: Id977cbbb75fd2a83f2eb6f4e53eb03ff18049b80
Jira: BENT-752