Allow configuration through config file
Add a config file which can be used to specify sensible defaults.
These will be used when the corresponding flag is not specified. If
a variable is not specified either as a flag or in the config file,
it is still possible that it will fallback to some internal default
- if there is no default, the program will exit prompting you to
specify a value.