Renamed example org.json and added .dockerignore
Renamed the example organization.json file so that
it is conistient with the example config.conf file.

Also added a .dockerignore file so that we never pull in
configuration files in the image. Also, this will speed up builds
since making changes to the config files, Makefiles or README no
longer invalidates the cache.