Cleanup function for keeping X latest components
This is a rough implementation of a new clenaup function which will
look at AMIs with the name structure <component>-<creation timestamp>
and only keep the X newest for each component.

Since this is a rough implementation, it might still need some
modifications to fit your needs better, but should definetly work.
It just needs to be added under the "mark-for-cleanup" case in the
main cloudsweeper package.